// src/gql/__generated__/product.ts
export type ColorsFragment = {
  id: string;
  name: string;
  color:
    | 'anthracite_oxygen'
    | 'berluti_edition'
    | 'black'
    | 'black_aluminium'
    | 'black_anthracite'
    | 'black_fabric'
    | 'brass_tone_aluminium'
    | 'bronze'
    | 'bronze_tone_aluminium'
    | 'bronze_walnut'
    | 'chestnut'
    | 'dark_maroon'
    | 'fabric'
    | 'gold'
    | 'gold_oak'
    | 'gold_tone'
    | 'green'
    | 'grey_mist'
    | 'infantry_green'
    | 'light_oak_wood'
    | 'natural'
    | 'natural_aluminium'
    | 'navy'
    | 'nordic_ice'
    | 'oak_wood'
    | 'oxygen_blue'
    | 'parisian_night_blue'
    | 'pink'
    | 'purple_heart'
    | 'sand'
    | 'silver'
    | 'silver_natural_oak'
    | 'timber'
    | 'white_marble';
  colorCode: {
    hex: any;
  };
  __typename: 'ProductColor';
};
export type SizesFragment = {
  id: string;
  name: string;
  size: 's_48' | 's_55' | 's_65' | 's_77' | 's_88';
  __typename: 'ProductSize';
};
export type PositionningsFragment = {
  id: string;
  name: string;
  positionning: 'fixed' | 'floor' | 'motorized' | 'table' | 'wall_mount';
  __typename: 'ProductPositionning';
};
export type FrameColorsFragment = {
  id: string;
  name: string;
  frameColor: 'black' | 'gold' | 'silver';
  colorCode: {
    hex: any;
  };
  __typename: 'ProductFrameColor';
};
export type SoundbarColorsFragment = {
  id: string;
  name: string;
  color: 'grey_melange' | 'light_oak' | 'smocked_oak' | 'walnut';
  colorCode: {
    hex: any;
  };
  __typename: 'ProductSoundbarColor';
};
export type SupportColorsFragment = {
  id: string;
  name: string;
  supportColor: 'bronze' | 'piano_black' | 'silver';
  colorCode: {
    hex: any;
  };
  __typename: 'ProductSupportColor';
};
export type ShopFragment = {
  id: string;
  name: string;
  email: string | null;
  address: string;
  city: string;
  postalCode: string;
  country: string;
  phoneNumber: string;
  googleMapsUrl: string | null;
  image: {
    id: string;
    url: string;
  } | null;
  openHours: any | null;
};
export type AssetsByProductVariantFragment = {
  id: string;
  asset: {
    url: string;
  }[];
  color:
    | (
        | 'anthracite_oxygen'
        | 'berluti_edition'
        | 'black'
        | 'black_aluminium'
        | 'black_anthracite'
        | 'black_fabric'
        | 'brass_tone_aluminium'
        | 'bronze'
        | 'bronze_tone_aluminium'
        | 'bronze_walnut'
        | 'chestnut'
        | 'dark_maroon'
        | 'fabric'
        | 'gold'
        | 'gold_oak'
        | 'gold_tone'
        | 'green'
        | 'grey_mist'
        | 'infantry_green'
        | 'light_oak_wood'
        | 'natural'
        | 'natural_aluminium'
        | 'navy'
        | 'nordic_ice'
        | 'oak_wood'
        | 'oxygen_blue'
        | 'parisian_night_blue'
        | 'pink'
        | 'purple_heart'
        | 'sand'
        | 'silver'
        | 'silver_natural_oak'
        | 'timber'
        | 'white_marble'
      )
    | null;
  positionning:
    | ('fixed' | 'floor' | 'motorized' | 'table' | 'wall_mount')
    | null;
  soundbarColor:
    | ('grey_melange' | 'light_oak' | 'smocked_oak' | 'walnut')
    | null;
  size: ('s_48' | 's_55' | 's_65' | 's_77' | 's_88') | null;
  frameColor: ('black' | 'gold' | 'silver') | null;
  supportColor: ('bronze' | 'piano_black' | 'silver') | null;
};
export type ProductFragment = {
  id: string;
  slug: string;
  name: string;
  description: string | null;
  price: number | null;
  mainAsset: {
    url: string;
  };
  assets: {
    url: string;
  }[];
  category: {
    name: string;
    slug: string;
  } | null;
  isNew: boolean;
  colors: ColorsFragment[];
  sizes: SizesFragment[];
  positionnings: PositionningsFragment[];
  frameColors: FrameColorsFragment[];
  soundbarColors: SoundbarColorsFragment[];
  supportColors: SupportColorsFragment[];
  quantity: number;
  shops: ShopFragment[];
  assetsByProductVariants: AssetsByProductVariantFragment[];
};
export type Product = {
  product: ProductFragment | null;
};